Add authorization, policies

+++ app/Http/Controllers/Admin/Users/AdminUsersController.php
@@ -6,6 +6,8 @@ use App;
 use App\Http\Controllers\Admin\AdminBaseController;
 use App\Http\Requests\AdminUsersStoreRequest;
 use App\Http\Requests\AdminUsersUpdateRequest;
+use App\Models\User;
+use App\Policies\Ability;
 use App\Services\UsersService;
 
 final class AdminUsersController extends AdminBaseController
@@ -21,10 +23,10 @@ final class AdminUsersController extends AdminBaseController
 
     public function index()
     {
+        $this->authorize(Ability::VIEW_ANY, User::class);
         $users = $this->usersService->paginateUsers(self::DEFAULT_MODELS_PER_PAGE);
         $statuses = $this->getStatuses();
         $roles = $this->getRoles();
-
         return view('admin.users.index', compact('users', 'statuses', 'roles'));
     }
 
@@ -40,6 +42,7 @@ final class AdminUsersController extends AdminBaseController
 
     public function store(AdminUsersStoreRequest $request)
     {
+        $this->authorize(Ability::STORE, User::class);
         $user = $this->usersService->createUser($request->getFormData());
 
         return response()->json($user);
@@ -47,6 +50,7 @@ final class AdminUsersController extends AdminBaseController
 
     public function show($id)
     {
+        $this->authorize(Ability::VIEW, User::class);
         $user = $this->usersService->findUser($id);
 
         return response()->json($user);
@@ -54,6 +58,7 @@ final class AdminUsersController extends AdminBaseController
 
     public function update($id, AdminUsersUpdateRequest $request)
     {
+        $this->authorize(Ability::UPDATE, User::class);
         $user = $this->usersService->updateUser($id, $request->getFormUpdateData());
 
         return response()->json($user);
@@ -61,6 +66,7 @@ final class AdminUsersController extends AdminBaseController
 
     public function destroy($id)
     {
+        $this->authorize(Ability::DELETE, User::class);
         $user = $this->usersService->deleteUser($id);
 
         return response()->json($user);
